MACAddr v0.0.1 MACAddr.Local

Convenience functions to determine a local machine’s MAC addresses, using Erlang’s inet module.

Functions

all()

Returns a list of all local MAC addresses.

Examples

Get a list of local MAC addresses:

iex> MACAddr.Local.all
[<<244, 92, 137, 11, 149, 14>>, <<74, 0, 5, 15, 252, 96>>,
 <<74, 0, 5, 7, 107, 15>>, <<6, 92, 137, 100, 224, 92>>,
 <<2, 127, 112, 79, 147, 123>>, <<246, 92, 137, 189, 168, 100>>]

Get a list of local MAC addresses, extract the universally administered ones, strip them to their OUIs, create random addresses from the OUIs, and convert them to Cisco-formatted strings:

iex> MACAddr.Local.all |> Enum.filter_map(fn(addr) ->
       MACAddr.is_universal?(addr)
     end, fn(addr) ->
       addr
         |> MACAddr.oui
         |> MACAddr.random
         |> MACAddr.format_as(:cisco)
     end)
["f45c.890b.950e"]

by_interface(name)

Returns the MAC address of the interface with name, or nil if it can’t find anything.

Examples

Get the MAC address of en0:

iex> MACAddr.Local.by_interface("en0")
<<244, 92, 137, 11, 149, 14>>

Try to get the MAC address of yeast0:

iex> MACAddr.Local.by_interface("yeast0")
nil

by_ip_address(ip_address)

Returns the MAC address of the interface with the IP address ip_address, or nil if it can’t find anything.

Examples

Get the MAC address associated with 192.168.20.1:

iex> MACAddr.Local.by_ip_address("192.168.20.1")
<<244, 92, 137, 11, 149, 14>>